+++ components/rightColumn/results.tsx
@@ -1,14 +1,14 @@
 import { BarChart3, CheckCircle, Link, RefreshCw } from "lucide-react";
-import { ResultsProps } from "@/types/propTypes";
+import { useAppStore } from "@/store/store";
 import SampleArticles from "@/components/rightColumn/sampleArticles";
 
-const Results = ({
-  isAnalyzing,
-  results,
-  inputType,
-  inputValue,
-  userContent,
-}: ResultsProps) => {
+const Results = () => {
+  const isAnalyzing = useAppStore((state) => state.isAnalyzing);
+  const results = useAppStore((state) => state.results);
+  const inputType = useAppStore((state) => state.inputType);
+  const inputValue = useAppStore((state) => state.inputValue);
+  const userContent = useAppStore((state) => state.userContent);
+
   return (
     <div className="[&::-webkit-scrollbar-thumb]:bg-transparent [&::-webkit-scrollbar-track]:bg-transparent hover:[&::-webkit-scrollbar-thumb]:bg-gray-300 [&::-webkit-scrollbar-thumb]:rounded-full [&::-webkit-scrollbar-track]:rounded-full [&::-webkit-scrollbar]:w-1 max-h-[82vh] overflow-y-scroll transition-all ease-in-out">
       <h2 className="flex items-center mb-2 font-medium text-gray-800 text-xl">
